industrial_iot/mqtt_bridge/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
httpx==0.25.2
pydantic==1.10.13
python-dotenv==1.0.0